+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 - Bumped OpenAI dependency to `>=1.108.1` to support their breaking change: "change optional parameter type from `NotGiven` to `Omit`"
+- `get_selected_backend_keys()` now correctly considers backends enabled by default (like before v0.15.0)
 
 ## [v0.15.0] - 2025-11-07